射撃処理記述開始

+++ src/main/java/types/GunData.java
@@ -19,14 +19,14 @@ public class GunData extends DataBase {
 	public int ADS_TICK = 0;
 	/** リロード ; リロードにtickかかる : int型 **/
 	public int RELOAD_TICK= 20;
-	/** レート ; レートtick間隔で発射する : int型 **/
-	public int RATE_TICK = 2;
+	/** レート ; レートtick間隔で発射する : float型 **/
+	public float RATE_TICK = 2;
 	/** 射撃モード : String配列型 **/
 	public String[] FIREMODE=new String[] { "semiauto" };
 	/** 貫通力 貫通力体のMOBにダメージが与えられる -1で∞ : int型 **/
 	public int BULLET_POWER = 1;
-	/** バーストのレート : int型 **/
-	public int BURST_RATE_TICK = 1;
+	/** バーストのレート : float型 **/
+	public float BURST_RATE_TICK = 1;
 	/** バーストの発射数 : int型 **/
 	public int BURST_BULLET_NUM = 3;
 	/** 装填数 : int型 **/
